move OTP methods into separate classes/modules
