+++ app/services/phone_confirmation/code_generator.rb
@@ -1,8 +1,10 @@
+require 'otp_code_generator'
+
 module PhoneConfirmation
   class CodeGenerator
     def self.call
       digits = TwoFactorAuthenticatable.direct_otp_length
-      SecureRandom.random_number(10**digits).to_s.rjust(digits, '0')
+      OtpCodeGenerator.generate_digits(digits)
     end
   end
 end
